+++ xen/include/asm-x86/hvm/vmx/vmcs.h
@@ -31,7 +31,7 @@ extern int  vmx_cpu_up(void);
 extern void vmx_cpu_down(void);
 extern void vmx_save_host_msrs(void);
 extern unsigned long long ple_count;
-extern int ple_table_size;
+extern unsigned long long ple_table_size;
 extern int ple_table_mode;
 extern int ple_sched;
